+++ packages/browser/src/integrations/breadcrumbs.ts
@@ -1,23 +1,23 @@
+// TODO: Rename this whole file to `instrument.ts` and make a distinction between instrumenting (wrapping the API)
+//       and creating a breadcrumb (writing an `InstrumentHandler`)
+
 import { API, getCurrentHub } from '@sentry/core';
-import { Breadcrumb, BreadcrumbHint, Integration, Severity, WrappedFunction } from '@sentry/types';
-import {
-  fill,
-  getEventDescription,
-  getGlobalObject,
-  isString,
-  logger,
-  normalize,
-  parseUrl,
-  safeJoin,
-  supportsHistory,
-  supportsNativeFetch,
-} from '@sentry/utils';
+import { Integration, WrappedFunction } from '@sentry/types';
+import { fill, getGlobalObject, isString, logger, supportsHistory, supportsNativeFetch } from '@sentry/utils';
 
 import { BrowserClient } from '../client';
 import { breadcrumbEventHandler, keypressEventHandler, wrap } from '../helpers';
 
+import {
+  defaultHandlers,
+  InstrumentHandler,
+  InstrumentHandlerCallback,
+  InstrumentHandlerType,
+} from './instrumenthandlers';
+
 const global = getGlobalObject<Window>();
 let lastHref: string | undefined;
+
 /**
  * @hidden
  */
@@ -38,11 +38,15 @@ interface BreadcrumbIntegrations {
   history?: boolean;
   sentry?: boolean;
   xhr?: boolean;
+  handlers?: InstrumentHandler[];
 }
 
 type XMLHttpRequestProp = 'onload' | 'onerror' | 'onprogress';
 
-/** Default Breadcrumbs instrumentations */
+/**
+ * Default Breadcrumbs instrumentations
+ * @deprecated With v6, this will be renamed
+ */
 export class Breadcrumbs implements Integration {
   /**
    * @inheritDoc
@@ -57,6 +61,16 @@ export class Breadcrumbs implements Integration {
   /** JSDoc */
   private readonly _options: BreadcrumbIntegrations;
 
+  /** JSDoc */
+  private readonly _handlers: { [key in InstrumentHandlerType]: InstrumentHandlerCallback[] } = {
+    console: [],
+    dom: [],
+    fetch: [],
+    history: [],
+    sentry: [],
+    xhr: [],
+  };
+
   /**
    * @inheritDoc
    */
@@ -70,6 +84,40 @@ export class Breadcrumbs implements Integration {
       xhr: true,
       ...options,
     };
+    this._setupHandlers([...defaultHandlers, ...(this._options.handlers || [])]);
+  }
+
+  /** JSDoc */
+  private _setupHandlers(handlers: InstrumentHandler[]): void {
+    for (const handler of handlers) {
+      // tslint:disable-next-line:strict-type-predicates
+      if (!handler || typeof handler.type !== 'string' || typeof handler.callback !== 'function') {
+        continue;
+      }
+      this._handlers[handler.type].push(handler.callback);
+    }
+  }
+
+  /** JSDoc */
+  private _triggerHandlers(type: InstrumentHandlerType, data: any): void {
+    if (!getCurrentHub().getIntegration(Breadcrumbs)) {
+      return;
+    }
+
+    if (!type || !this._handlers[type]) {
+      return;
+    }
+
+    for (const handler of this._handlers[type]) {
+      try {
+        handler(data);
+      } catch (e) {
+        logger.error(
+          `Error while triggering instrumentation handler.\nType: ${type}\nName: ${handler.name ||
+            '<anonymous>'}\nError: ${e}`,
+        );
+      }
+    }
   }
 
   /** JSDoc */
@@ -77,6 +125,9 @@ export class Breadcrumbs implements Integration {
     if (!('console' in global)) {
       return;
     }
+
+    const triggerHandlers = this._triggerHandlers.bind(this, 'console');
+
     ['debug', 'info', 'warn', 'error', 'log', 'assert'].forEach(function(level: string): void {
       if (!(level in global.console)) {
         return;
@@ -84,33 +135,11 @@ export class Breadcrumbs implements Integration {
 
       fill(global.console, level, function(originalConsoleLevel: () => any): Function {
         return function(...args: any[]): void {
-          const breadcrumbData = {
-            category: 'console',
-            data: {
-              extra: {
-                arguments: normalize(args, 3),
-              },
-              logger: 'console',
-            },
-            level: Severity.fromString(level),
-            message: safeJoin(args, ' '),
+          const handlerData = {
+            args,
+            level,
           };
-
-          if (level === 'assert') {
-            if (args[0] === false) {
-              breadcrumbData.message = `Assertion failed: ${safeJoin(args.slice(1), ' ') || 'console.assert'}`;
-              breadcrumbData.data.extra.arguments = normalize(args.slice(1), 3);
-              Breadcrumbs.addBreadcrumb(breadcrumbData, {
-                input: args,
-                level,
-              });
-            }
-          } else {
-            Breadcrumbs.addBreadcrumb(breadcrumbData, {
-              input: args,
-              level,
-            });
-          }
+          triggerHandlers(handlerData);
 
           // this fails for some browsers. :(
           if (originalConsoleLevel) {
@@ -215,83 +244,35 @@ export class Breadcrumbs implements Integration {
       return;
     }
 
+    const triggerHandlers = this._triggerHandlers.bind(this, 'fetch');
+
     fill(global, 'fetch', function(originalFetch: () => void): () => void {
       return function(...args: any[]): void {
-        const fetchInput = args[0];
-        let method = 'GET';
-        let url;
-
-        if (typeof fetchInput === 'string') {
-          url = fetchInput;
-        } else if ('Request' in global && fetchInput instanceof Request) {
-          url = fetchInput.url;
-          if (fetchInput.method) {
-            method = fetchInput.method;
-          }
-        } else {
-          url = String(fetchInput);
-        }
-
-        if (args[1] && args[1].method) {
-          method = args[1].method;
-        }
-
-        const client = getCurrentHub().getClient<BrowserClient>();
-        const dsn = client && client.getDsn();
-        if (dsn) {
-          const filterUrl = new API(dsn).getStoreEndpoint();
-          // if Sentry key appears in URL, don't capture it as a request
-          // but rather as our own 'sentry' type breadcrumb
-          if (filterUrl && url.indexOf(filterUrl) !== -1) {
-            if (method === 'POST' && args[1] && args[1].body) {
-              addSentryBreadcrumb(args[1].body);
-            }
-            return originalFetch.apply(global, args);
-          }
-        }
-
-        const fetchData: {
-          method: string;
-          url: string;
-          status_code?: number;
-        } = {
-          method: isString(method) ? method.toUpperCase() : method,
-          url,
+        const handlerData: { [key: string]: any } = {
+          args,
+          endTimestamp: Date.now(),
+          fetchData: {
+            method: getFetchMethod(args),
+            url: getFetchUrl(args),
+          },
+          startTimestamp: Date.now(),
         };
 
-        return originalFetch
-          .apply(global, args)
-          .then((response: Response) => {
-            fetchData.status_code = response.status;
-            Breadcrumbs.addBreadcrumb(
-              {
-                category: 'fetch',
-                data: fetchData,
-                type: 'http',
-              },
-              {
-                input: args,
-                response,
-              },
-            );
+        return originalFetch.apply(global, args).then(
+          (response: Response) => {
+            handlerData.endTimestamp = Date.now();
+            handlerData.response = response;
+            handlerData.fetchData.status_code = response.status;
+            triggerHandlers(handlerData);
             return response;
-          })
-          .then(null, (error: Error) => {
-            Breadcrumbs.addBreadcrumb(
-              {
-                category: 'fetch',
-                data: fetchData,
-                level: Severity.Error,
-                type: 'http',
-              },
-              {
-                error,
-                input: args,
-              },
-            );
-
+          },
+          (error: Error) => {
+            handlerData.endTimestamp = Date.now();
+            handlerData.error = error;
+            triggerHandlers(handlerData);
             throw error;
-          });
+          },
+        );
       };
     });
   }
@@ -302,63 +283,37 @@ export class Breadcrumbs implements Integration {
       return;
     }
 
-    const captureUrlChange = (from: string | undefined, to: string | undefined): void => {
-      const parsedLoc = parseUrl(global.location.href);
-      const parsedTo = parseUrl(to as string);
-      let parsedFrom = parseUrl(from as string);
+    const triggerHandlers = this._triggerHandlers.bind(this, 'history');
 
-      // Initial pushState doesn't provide `from` information
-      if (!parsedFrom.path) {
-        parsedFrom = parsedLoc;
-      }
-
-      // because onpopstate only tells you the "new" (to) value of location.href, and
-      // not the previous (from) value, we need to track the value of the current URL
-      // state ourselves
-      lastHref = to;
-
-      // Use only the path component of the URL if the URL matches the current
-      // document (almost all the time when using pushState)
-      if (parsedLoc.protocol === parsedTo.protocol && parsedLoc.host === parsedTo.host) {
-        // tslint:disable-next-line:no-parameter-reassignment
-        to = parsedTo.relative;
-      }
-      if (parsedLoc.protocol === parsedFrom.protocol && parsedLoc.host === parsedFrom.host) {
-        // tslint:disable-next-line:no-parameter-reassignment
-        from = parsedFrom.relative;
-      }
-
-      Breadcrumbs.addBreadcrumb({
-        category: 'navigation',
-        data: {
-          from,
-          to,
-        },
-      });
-    };
-
-    // record navigation (URL) changes
     const oldOnPopState = global.onpopstate;
     global.onpopstate = (...args: any[]) => {
-      const currentHref = global.location.href;
-      captureUrlChange(lastHref, currentHref);
+      const to = global.location.href;
+      const handlerData = {
+        from: lastHref,
+        to,
+      };
+      // keep track of the current URL state, as we always receive only the updated state
+      lastHref = to;
+      triggerHandlers(handlerData);
       if (oldOnPopState) {
         return oldOnPopState.apply(this, args);
       }
     };
 
-    /**
-     * @hidden
-     */
+    /** @hidden */
     function historyReplacementFunction(originalHistoryFunction: () => void): () => void {
-      // note history.pushState.length is 0; intentionally not declaring
-      // params to preserve 0 arity
       return function(this: History, ...args: any[]): void {
         const url = args.length > 2 ? args[2] : undefined;
-        // url argument is optional
         if (url) {
           // coerce to string (this is what pushState does)
-          captureUrlChange(lastHref, String(url));
+          const to = String(url);
+          const handlerData = {
+            from: lastHref,
+            to,
+          };
+          // keep track of the current URL state, as we always receive only the updated state
+          lastHref = to;
+          triggerHandlers(handlerData);
         }
         return originalHistoryFunction.apply(this, args);
       };
@@ -373,6 +328,8 @@ export class Breadcrumbs implements Integration {
     if (!('XMLHttpRequest' in global)) {
       return;
     }
+
+    const triggerHandlers = this._triggerHandlers.bind(this, 'xhr');
 
     /**
      * @hidden
@@ -427,19 +384,20 @@ export class Breadcrumbs implements Integration {
       originalSend =>
         function(this: SentryWrappedXMLHttpRequest, ...args: any[]): void {
           const xhr = this; // tslint:disable-line:no-this-assignment
+          const handlerData: { [key: string]: any } = {
+            args,
+            endTimestamp: Date.now(),
+            startTimestamp: Date.now(),
+            xhr,
+          };
 
-          if (xhr.__sentry_own_request__) {
-            addSentryBreadcrumb(args[0]);
-          }
+          triggerHandlers(handlerData);
 
           /**
            * @hidden
            */
           function onreadystatechangeHandler(): void {
             if (xhr.readyState === 4) {
-              if (xhr.__sentry_own_request__) {
-                return;
-              }
               try {
                 // touching statusCode in some platforms throws
                 // an exception
@@ -449,16 +407,9 @@ export class Breadcrumbs implements Integration {
               } catch (e) {
                 /* do nothing */
               }
-              Breadcrumbs.addBreadcrumb(
-                {
-                  category: 'xhr',
-                  data: xhr.__sentry_xhr__,
-                  type: 'http',
-                },
-                {
-                  xhr,
-                },
-              );
+              handlerData.endTimestamp = Date.now();
+              handlerData.requestComplete = true;
+              triggerHandlers(handlerData);
             }
           }
 
@@ -495,17 +446,6 @@ export class Breadcrumbs implements Integration {
   }
 
   /**
-   * Helper that checks if integration is enabled on the client.
-   * @param breadcrumb Breadcrumb
-   * @param hint BreadcrumbHint
-   */
-  public static addBreadcrumb(breadcrumb: Breadcrumb, hint?: BreadcrumbHint): void {
-    if (getCurrentHub().getIntegration(Breadcrumbs)) {
-      getCurrentHub().addBreadcrumb(breadcrumb, hint);
-    }
-  }
-
-  /**
    * Instrument browser built-ins w/ breadcrumb capturing
    *  - Console API
    *  - DOM API (click/typing)
@@ -532,23 +472,24 @@ export class Breadcrumbs implements Integration {
   }
 }
 
-/** JSDoc */
-function addSentryBreadcrumb(serializedData: string): void {
-  // There's always something that can go wrong with deserialization...
-  try {
-    const event = JSON.parse(serializedData);
-    Breadcrumbs.addBreadcrumb(
-      {
-        category: 'sentry',
-        event_id: event.event_id,
-        level: event.level || Severity.fromString('error'),
-        message: getEventDescription(event),
-      },
-      {
-        event,
-      },
-    );
-  } catch (_oO) {
-    logger.error('Error while adding sentry type breadcrumb');
+/** Extract `method` from fetch call arguments */
+function getFetchMethod(fetchArgs: any[] = []): string {
+  if ('Request' in global && fetchArgs[0] instanceof Request && fetchArgs[0].method) {
+    return String(fetchArgs[0].method).toUpperCase();
   }
+  if (fetchArgs[1] && fetchArgs[1].method) {
+    return String(fetchArgs[1].method).toUpperCase();
+  }
+  return 'GET';
+}
+
+/** Extract `url` from fetch call arguments */
+function getFetchUrl(fetchArgs: any[] = []): string {
+  if (typeof fetchArgs[0] === 'string') {
+    return fetchArgs[0];
+  }
+  if ('Request' in global && fetchArgs[0] instanceof Request) {
+    return fetchArgs[0].url;
+  }
+  return String(fetchArgs[0]);
 }